+++ R/bilat.symmetry.r
@@ -261,7 +261,7 @@ bilat.symmetry <- function(A, ind = NULL, side = NULL, replicate = NULL, object.
     ind <- factor(rep(ind,2)); side <- gl(2,n); if(!is.null(replicate)) replicate <- rep(replicate,2)
     if(print.progress) cat("\nObject Symmetry GPA\n")
     gpa.res <- gpagen(A, print.progress = print.progress)
-    A <- gpa.res$coords
+    A <- gpa.res$coords  
   }
   Y <- two.d.array(A)
   
@@ -344,9 +344,7 @@ bilat.symmetry <- function(A, ind = NULL, side = NULL, replicate = NULL, object.
   # build shape components for output
   X.ind <- model.matrix(~ind + 0, data = as.data.frame(dat.shape[-1]))
   symm.component <- arrayspecs(coef(lm.fit(X.ind, Y)),p,k)
-  ind.names <- substr(dimnames(symm.component)[[3]], start=4,
-                      stop=nchar(dimnames(symm.component)[[3]]))
-  dimnames(symm.component)[[3]] <- ind.names
+  dimnames(symm.component)[[3]] <- spec.names 
   X.side <- model.matrix(~(side:ind) + 0, data = as.data.frame(dat.shape[-1]))
   avg.side.symm <- coef(lm.fit(X.side, Y))
   n.ind <- nlevels(ind)
@@ -357,7 +355,7 @@ bilat.symmetry <- function(A, ind = NULL, side = NULL, replicate = NULL, object.
   asymm.component <- simplify2array(lapply(1:n.ind, function(j) {
     t(matrix(asymm.component[j,],k,p)) + mn.shape
     }))
-  dimnames(asymm.component)[[3]] <- ind.names
+  dimnames(asymm.component)[[3]] <- spec.names 
   
   DA.est <- coef(.lm.fit(X.side, Y))
   DA.mns <- arrayspecs(rbind(apply(DA.est[-indsq,], 2, mean), apply(DA.est[indsq,], 2, mean)), p, k)
@@ -368,12 +366,12 @@ bilat.symmetry <- function(A, ind = NULL, side = NULL, replicate = NULL, object.
   FA.component <- ind.mns[-indsq,] - ind.mns[indsq,]
   FA.component <- simplify2array(lapply(1:n.ind, function(j) 
   {t(matrix(FA.component[j,],k,p)) + mn.shape - mn.DA}))
-  dimnames(FA.component)[[3]] <- ind.names 
+  dimnames(FA.component)[[3]] <- spec.names  
   
   # Calculate individual asymmetry indices
   signed.asymm <- two.d.array(asymm.component)
   signed.AI <- sqrt(apply(signed.asymm^2, 1, sum))
-  names(signed.AI) <- ind.names
+  names(signed.AI) <- spec.names 
   
   asymm.mean <- apply(signed.asymm, 2, mean)
   unsigned.asymm <- matrix(NA, nrow=nrow(signed.asymm), ncol=ncol(signed.asymm))
@@ -381,7 +379,7 @@ bilat.symmetry <- function(A, ind = NULL, side = NULL, replicate = NULL, object.
     unsigned.asymm[i,] <- ifelse(signed.asymm[i,]%*%asymm.mean > 0, signed.asymm[i,], signed.asymm[i,]*(-1))
   } 
   unsigned.AI <- sqrt(apply(unsigned.asymm^2, 1, sum))
-  names(unsigned.AI) <- ind.names
+  names(unsigned.AI) <- spec.names 
 
   out <- list(shape.anova = shape.anova, symm.shape = symm.component,
               asymm.shape = asymm.component, DA.component = DA.mns, FA.component = FA.component,