+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Bug fixes
 * Fixed a bug in `logLikh` internal function that required `matrix` class objects for calculation.
 * Fixed a typo bug in `procD.lm` that caused lost results.
+* Fixed names being dropped from shape components in `bilat.symmetry`
 
 ### OTHER CHANGES
 * Additional error checking in `modularity.test` and `phylo.modularity` `rgl`